BUG 4.4
Extracted from the installer, this way, we start exactly from what most BUG users know. Should correspond to revision 2221 at SourceForge:
https://sourceforge.net/p/civ4bug/code/2221/

Folder structure set up for an installation in the Mods folder. That's the only way of installing that I intend to (fully) support. Also added Readme.txt and GameSetupCheck.txt from SVN - just to make clear that these are not post-4.4 additions. I'll probably replace the Readme - and remove GameSetupCheck and other parts of BUFFY in the end. There is only one HoF mod and this won't be it.