Update BULL to SVN revision r220
Obtained from
https://sourceforge.net/p/civ4bull/code/220/

The revision after that only adds subversion meta data; we don't want that. So this is mainly an update to v1.6 of the Unofficial Patch. I've omitted a change to the release date in the changelog - 20 Sept 2010 was the closest thing to an actual release; the GameCore DLL was not updated afterwards. (Re-compiling the DLL will be our next objective.)