Copies of Fractal, Pangaea - to get them listed on top.
I think it was a bad idea to bundle map scripts with BUG, but I think some of the modified BtS/ WL scripts aren't available as a separate download, so I suppose we're stuck with them. Let's then at least bump those funky map scripts off the top of the list by adding copies of Fractal and Pangaea from WL.

The PrivateMaps folder takes priority over PublicMaps when it comes to the order on the game setup screens. (Apart from that, I think the idea is that modified BtS scripts should go in PublicMaps and community-created scripts in PrivateMaps. But it doesn't really matter.)