Merge branch 'balloon'
That is, add a small framework for modifying the EXE at runtime, use that framework to decrease the size of the balloon icons (plot indicators, bubbles), add files where Taurus can add options and carve out space on the Mod Options screen, add options for the balloon icon size. And a tweak to make XML loading a little easier to debug. DLL updated (Release build).