Attempt at storing extra data in saves w/o breaking BtS
I.e. while maintaining bidirectional savegame compatibility with BtS. An initial test had suggested that the EXE will tolerate data appended at the end of the stream. That might be true, but I'm assuming now that the write function of CvStatistics isn't actually the end of the stream and that there is some sort of coda written by the EXE.

Not worth further effort; I don't think I'll need to save any extra data anyway, and, if I do, some information can be squeezed into the medium and high bits of the uiFlag variables.