Handling of Locked Assets, BUFFY
BUFFY saves can't be loaded; I've realized that they're encrypted. (And I'm not going to hunt for the key in the binary code of the BUFFY DLL.)

The Locked Assets check was actually difficult to bypass, but I think I've gotten it to work now. Still want to add a findAddressOffset call for robustness (tbd.).

HTML help updated.