New tab "Saves" on mod options screen
I've moved the Start/End Save options there, making room on the General tab, the Taurus mod name and replay name options, previously on the System tab, and the MapFinder options from the Map tab. That's a bit of a stretch, but the MapFinder does create savegames. Also made clear in hover text and help file that MapFinder is not bundled with Taurus, needs to be installed separately.

I'll add at least two more savegame-related options to the Saves tab. The freed space on the General tab will be used for options that disable the flashing hints on the main screen and city screen. I've already moved the option for the blinking Ctrl+Alt+O hint there, and, now first of all on the upper left, the option for showing or hiding the mod options button. Both moved from the System tab.

Help file updated, reverting some accidental replacements of translations with English text.