Remove Smart_Map to reduce memory consumption
Inspired by this C2C Git commit:
https://github.com/caveman2cosmos/Caveman2Cosmos/commit/1a0d1c6e32ebb2b1fd93964d1cde66ec511a75df

In a test on my PC, the memory use of the BtS process while on the Custom Game screen was 250 MB with Smart_Map present (but without having actually selected it) and only 150 MB without it. It sounds like those 100 MB will remain in use once a game has been launched (too lazy to double-check that), and I doubt that it even matters whether a game setup screen has been opened. Ideally, the memory allocations of the script would be somehow delayed until the player selects it. That should be doable, but I don't want to investigate this. 100 MB is of course not a big deal nowadays; I don't suppose that it matters for power consumption either. Still, I feel that the problem should be flagged somehow, and letting the (rare) player who actually uses the script install it manually will accomplish that. Imo map scripts should've never been bundled with BUG anyway.