Proper sea level option for Big and Small map script
The option previously was enabled but had no effect. I've aimed at about the same effect on the water percentage as on Fractal, i.e. +6 points at High, -8 points at Low, but it's not obvious that this is the best approach for this rather land-heavy map script - at Normal sea level, the water percentage is already close to 30%, which corresponds more closely to Low sea level on Fractal.

Based on this AdvCiv commit:
https://github.com/f1rpo/AdvCiv/commit/d6df04de8682f7555fde7251d8a2304f1e558454

Related CFC post:
https://forums.civfanatics.com/threads/taurus.677586/post-16616358