Fix early memory allocation by SmartMap
Related CFC post:
https://forums.civfanatics.com/threads/taurus.677586/post-16734603

By defining the static data for Earth maps as strings (compressed strings even) and only converting them to int arrays if and when that data is needed. The main issue seems to be the Python overhead for thousands of int objects. I'm not sure if the deallocation of those objects should be (or is in fact) problematic. Just avoiding costly allocations upon loading of the module is already an improvement, in any case. No problems, that way, for players who don't use Smart Map.