Allow Custom Assets in ini file
I'd really like to allow only art to be loaded from CustomAssets, but I've come to realize that this isn't possible. (Loading everything -except- art can be done.) So, to allow Blue Marble to be loaded from CustomAssets, which the Taurus install instructions say is possible, all CustomAssets will have to be allowed. I guess XML and Python changes - if they make sense in CustomAssets at all - do have a decent chance of being compatible with Taurus. So maybe this change is all-upside. It's also the default mod behavior.