Stop 1st-contact messages from indicating unobservable units
When a rival unit moves away right after a first-contact meeting and before the main map gets updated graphically, the player has no chance to observe the unit that causes the meeting, and, in that case, the optional on-screen message about the meeting also shouldn't indicate where the meeting occurred. (Even though it's confusing to be informed of a meeting - or even having the meeting via the diplo screen - and remaining unaware of its location.)

I suppose the changes in CvTeam::makeHasMet are only refactoring or future-proofing.