Remove Fractal, Pangaea copies
The idea was for those two to appear at the top of the drop-down menu. I'll be deploying the (meaningfully modified or new) map scripts in a separate zip archive in the future (though they'll stay unzipped on Git), so, unless players manually unpack those, there will no longer be a problem of modified and new map scripts at the top of the menu making normal scripts like Fractal and Pangaea difficult to find. For players who do unpack the zip archive, while pushing Fractal and Pangaea to the top would be useful, I don't want those players to have to deal with two different Maps folders; it could be confusing.

To clarify: The order of the maps menu can't (easily) be modified. Maps included with the mod are listed first and PrivateMaps before PublicMaps.