Rename PublicMaps to PrivateMaps
Now that there'll be only one Maps folder, PrivateMaps is, I think, the better choice. I don't think maps that don't exist in BtS are supposed to be in a PublicMaps folder (because they'd get written into CivilizationIV.ini? not sure). And I only want there to be one Maps folder (cf. the parent commit's message).